web/requirements.txt
Flask==3.0.0
authlib==1.2.1
Werkzeug==3.0.1
Pillow==10.1.0
tensorflow==2.15.0
numpy==1.24.3

